Refactor filter properties into Filters context
